Add IterativeComb and SmartBogo to categories for RunAllSorts
Length 11 on SmartBogoSort gives a good visual without being boringly long, which would be useful for video makers. It breaks the pattern of powers of 2 however, but there are currently no Delays built into the sort that would allow this to be manipulated using a speed modification instead. I put 8 for the length even though it's instant because 16 takes an obscene amount of time